+++ src/webview/WebviewProvider/lib/platform/utils/index.ts
@@ -7,4 +7,4 @@ export { default as getRepoInfo } from "./getRepoInfo";
 export { default as fetchPipelines } from "./fetchPipelines";
 export { default as fetchDatasets } from "./fetchDatasets";
 export { default as fetchDataLinks } from "./fetchDataLinks";
-export { default as generateTest } from "./generateTest";
+export { default as createTest } from "./createTest";
